requirements/base.txt
Django==1.6.5
bpython==0.13.1
django-braces==1.4.0
django-model-utils==2.2
logutils==0.3.3
South==1.0
psycopg2
django-tastypie
jsonfield
django-encrypted-fields
requests
PILLOW
django-cors-headers
pyquery
django-taggit
sorl-thumbnail
